refactor: 下沉 jjc_ranking_service 单例并解耦 status_monitor
